Check the patch file is valid before applying patches in postinstall
