+++ scripts/apply_patches.sh
@@ -3,6 +3,8 @@
 # Script to apply the patches in patches directory, using patch-package.
 set -e
 
+node scripts/checkPatchFile.js
+
 for d in patches/*/ ; do
     echo "Patching $d..."
     yarn patch-package --patch-dir "$d"
